requirements.txt
matplotlib
pytorch
torchvision
tqdm
wandb
